Add likelihood calculation using tensor tools.
This give a much faster and easier to understand approach, but I have not yet removed the old approach.
I have added tests to verify that the calculation is matching for the likelihood with and without partial observations and different choices of fitting times.
However, there are other variables being calcuated usign the other approach, and it is not clear which are needed and where. These are left in pace for now until we verify that all calculations and data storage is correct and as needed for plotting functions.